pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verified-koopman"
version = "0.1.0"
description = "Proof-carrying nucleus-bottleneck Koopman autoencoders + verified Heyting structure (Lean 4)."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Verified Koopman Authors" }]
dependencies = [
  "numpy>=1.23",
  "scipy>=1.9",
  "torch>=1.10",
  "matplotlib>=3.5",
  "pyyaml>=6.0",
  "cvxpy>=1.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
